email-rag-4/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pst-email-rag-bot"
version = "0.1.0"
description = "AI-powered email analysis system for Outlook PST files using RAG technology"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "PST RAG Team" }
]
keywords = ["email", "pst", "rag", "llm", "fastapi", "forensics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # FastAPI & Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "gunicorn>=21.2.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",

    # Database - PostgreSQL
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",

    # Pydantic & Settings
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",

    # Redis & Caching
    "redis>=5.0.1",
    "aioredis>=2.0.1",

    # Celery & Task Queue
    "celery[redis]>=5.3.6",

    # Vector Database
    "chromadb>=0.4.22",

    # Embeddings
    "sentence-transformers>=2.2.2",
    "torch>=2.1.0",

    # LLM Providers
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.2",
    "langchain>=0.1.4",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.1",
    "langchain-google-genai>=0.0.6",
    "tiktoken>=0.5.2",

    # PST Parsing
    "libpff-python>=20231205",

    # Document Processing
    "pypdf>=3.17.4",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.2",
    "python-magic>=0.4.27",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",

    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.1.2,<5.0.0",

    # Logging
    "loguru>=0.7.2",

    # HTTP Client
    "httpx>=0.26.0",
    "aiohttp>=3.9.3",

    # Utilities
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
    "email-validator>=2.1.0",
    "orjson>=3.9.12",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",
    "faker>=22.2.0",

    # Linting & Formatting
    "ruff>=0.1.13",
    "black>=24.1.0",
    "isort>=5.13.2",
    "mypy>=1.8.0",

    # Type stubs
    "types-redis>=4.6.0",
    "types-passlib>=1.7.7",
    "types-python-jose>=3.3.4",
    "types-python-dateutil>=2.8.19",
]

docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.4",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
pstrag = "app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=term-missing --cov-report=html"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hatch
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | alembic
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["app"]
skip = ["alembic"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "chromadb.*",
    "sentence_transformers.*",
    "libpff.*",
    "pypff.*",
    "magic.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["app"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/alembic/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
